move spine to a separate file
