+++ packages/webgal/src/Core/controller/stage/pixi/PixiController.ts
@@ -8,9 +8,8 @@ import { logger } from '@/Core/util/logger';
 import { isIOS } from '@/Core/initializeScript';
 import { WebGALPixiContainer } from '@/Core/controller/stage/pixi/WebGALPixiContainer';
 import { WebGAL } from '@/Core/WebGAL';
-import 'pixi-spine'; // Do this once at the very start of your code. This registers the loader!
-import { Spine } from 'pixi-spine';
 import { SCREEN_CONSTANTS } from '@/Core/util/constants';
+import { addSpineBgImpl, addSpineFigureImpl } from '@/Core/controller/stage/pixi/spine';
 // import { figureCash } from '@/Core/gameScripts/vocal/conentsCash'; // 如果要使用 Live2D，取消这里的注释
 // import { Live2DModel, SoundManager } from 'pixi-live2d-display-webgal'; // 如果要使用 Live2D，取消这里的注释
 
@@ -78,21 +77,28 @@ export default class PixiStage {
   public readonly effectsContainer: PIXI.Container;
   public frameDuration = 16.67;
   public notUpdateBacklogEffects = false;
-  private readonly figureContainer: PIXI.Container;
-  private figureObjects: Array<IStageObject> = [];
-  private readonly backgroundContainer: PIXI.Container;
-  private backgroundObjects: Array<IStageObject> = [];
-
+  public readonly figureContainer: PIXI.Container;
+  public figureObjects: Array<IStageObject> = [];
+  public stageWidth = SCREEN_CONSTANTS.width;
+  public stageHeight = SCREEN_CONSTANTS.height;
+  public assetLoader = new PIXI.Loader();
+  public readonly backgroundContainer: PIXI.Container;
+  public backgroundObjects: Array<IStageObject> = [];
+  /**
+   * 添加 Spine 立绘
+   * @param key 立绘的标识，一般和立绘位置有关
+   * @param url 立绘图片url
+   * @param presetPosition
+   */
+  public addSpineFigure = addSpineFigureImpl.bind(this);
+  public addSpineBg = addSpineBgImpl.bind(this);
   // 注册到 Ticker 上的函数
   private stageAnimations: Array<IStageAnimationObject> = [];
-  private assetLoader = new PIXI.Loader();
   private loadQueue: { url: string; callback: () => void; name?: string }[] = [];
   private live2dFigureRecorder: Array<ILive2DRecord> = [];
-
   // 锁定变换对象（对象可能正在执行动画，不能应用变换）
   private lockTransformTarget: Array<string> = [];
-  private stageWidth = SCREEN_CONSTANTS.width;
-  private stageHeight = SCREEN_CONSTANTS.height;
+
   /**
    * 暂时没用上，以后可能用
    * @private
@@ -407,84 +413,6 @@ export default class PixiStage {
     }
   }
 
-  public addSpineBg(key: string, url: string) {
-    const spineId = `spine-${url}`;
-    const loader = this.assetLoader;
-    // 准备用于存放这个背景的 Container
-    const thisBgContainer = new WebGALPixiContainer();
-
-    // 是否有相同 key 的背景
-    const setBgIndex = this.backgroundObjects.findIndex((e) => e.key === key);
-    const isBgSet = setBgIndex >= 0;
-
-    // 已经有一个这个 key 的背景存在了
-    if (isBgSet) {
-      // 挤占
-      this.removeStageObjectByKey(key);
-    }
-
-    // 挂载
-    this.backgroundContainer.addChild(thisBgContainer);
-    const bgUuid = uuid();
-    this.backgroundObjects.push({
-      uuid: bgUuid,
-      key: key,
-      pixiContainer: thisBgContainer,
-      sourceUrl: url,
-      sourceType: 'live2d',
-      sourceExt: this.getExtName(url),
-    });
-
-    // 完成图片加载后执行的函数
-    const setup = () => {
-      const spineResource: any = this.assetLoader.resources?.[spineId];
-      // TODO：找一个更好的解法，现在的解法是无论是否复用原来的资源，都设置一个延时以让动画工作正常！
-      setTimeout(() => {
-        if (spineResource && this.getStageObjByUuid(bgUuid)) {
-          const bgSpine = new Spine(spineResource.spineData);
-          const transY = spineResource?.spineData?.y ?? 0;
-          /**
-           * 重设大小
-           */
-          const originalWidth = bgSpine.width; // TODO: 视图大小可能小于画布大小，应提供参数指定视图大小
-          const originalHeight = bgSpine.height; // TODO: 视图大小可能小于画布大小，应提供参数指定视图大小
-          const scaleX = this.stageWidth / originalWidth;
-          const scaleY = this.stageHeight / originalHeight;
-          logger.debug('bgSpine state', bgSpine.state);
-          // TODO: 也许应该使用 setAnimation 播放初始动画
-          if (bgSpine.spineData.animations.length > 0) {
-            // 播放首个动画
-            bgSpine.state.setAnimation(0, bgSpine.spineData.animations[0].name, true);
-          }
-          const targetScale = Math.max(scaleX, scaleY);
-          const bgSprite = new PIXI.Sprite();
-          bgSprite.addChild(bgSpine);
-          bgSprite.scale.x = targetScale;
-          bgSprite.scale.y = targetScale;
-          bgSprite.anchor.set(0.5);
-          bgSprite.position.y = this.stageHeight / 2;
-          thisBgContainer.setBaseX(this.stageWidth / 2);
-          thisBgContainer.setBaseY(this.stageHeight / 2);
-          thisBgContainer.pivot.set(0, this.stageHeight / 2);
-
-          // 挂载
-          thisBgContainer.addChild(bgSprite);
-        }
-      }, 0);
-    };
-
-    /**
-     * 加载器部分
-     */
-    this.cacheGC();
-    if (!loader.resources?.[url]) {
-      this.loadAsset(url, setup, spineId);
-    } else {
-      // 复用
-      setup();
-    }
-  }
-
   /**
    * 添加立绘
    * @param key 立绘的标识，一般和立绘位置有关
@@ -569,105 +497,6 @@ export default class PixiStage {
     this.cacheGC();
     if (!loader.resources?.[url]?.texture) {
       this.loadAsset(url, setup);
-    } else {
-      // 复用
-      setup();
-    }
-  }
-
-  /**
-   * 添加 Spine 立绘
-   * @param key 立绘的标识，一般和立绘位置有关
-   * @param url 立绘图片url
-   * @param presetPosition
-   */
-  public addSpineFigure(key: string, url: string, presetPosition: 'left' | 'center' | 'right' = 'center') {
-    const spineId = `spine-${url}`;
-    const loader = this.assetLoader;
-    // 准备用于存放这个立绘的 Container
-    const thisFigureContainer = new WebGALPixiContainer();
-
-    // 是否有相同 key 的立绘
-    const setFigIndex = this.figureObjects.findIndex((e) => e.key === key);
-    const isFigSet = setFigIndex >= 0;
-
-    // 已经有一个这个 key 的立绘存在了
-    if (isFigSet) {
-      this.removeStageObjectByKey(key);
-    }
-
-    const metadata = this.getFigureMetadataByKey(key);
-    if (metadata) {
-      if (metadata.zIndex) {
-        thisFigureContainer.zIndex = metadata.zIndex;
-      }
-    }
-    // 挂载
-    this.figureContainer.addChild(thisFigureContainer);
-    const figureUuid = uuid();
-    this.figureObjects.push({
-      uuid: figureUuid,
-      key: key,
-      pixiContainer: thisFigureContainer,
-      sourceUrl: url,
-      sourceType: 'live2d',
-      sourceExt: this.getExtName(url),
-    });
-
-    // 完成图片加载后执行的函数
-    const setup = () => {
-      const spineResource: any = this.assetLoader.resources?.[spineId];
-      // TODO：找一个更好的解法，现在的解法是无论是否复用原来的资源，都设置一个延时以让动画工作正常！
-      setTimeout(() => {
-        if (spineResource && this.getStageObjByUuid(figureUuid)) {
-          const figureSpine = new Spine(spineResource.spineData);
-          const spineBounds = figureSpine.getLocalBounds();
-          const spineCenterX = spineBounds.x + spineBounds.width / 2;
-          const spineCenterY = spineBounds.y + spineBounds.height / 2;
-          figureSpine.pivot.set(spineCenterX, spineCenterY);
-          // TODO: set animation 还没做
-          // figureSpine.state.setAnimation()
-          /**
-           * 重设大小
-           */
-          const originalWidth = figureSpine.width;
-          const originalHeight = figureSpine.height;
-          const scaleX = this.stageWidth / originalWidth;
-          const scaleY = this.stageHeight / originalHeight;
-          const targetScale = Math.min(scaleX, scaleY);
-          const figureSprite = new PIXI.Sprite();
-          figureSprite.addChild(figureSpine);
-          figureSprite.scale.x = targetScale;
-          figureSprite.scale.y = targetScale;
-          figureSprite.anchor.set(0.5);
-          figureSprite.position.y = this.stageHeight / 2;
-          const targetWidth = originalWidth * targetScale;
-          const targetHeight = originalHeight * targetScale;
-          thisFigureContainer.setBaseY(this.stageHeight / 2);
-          if (targetHeight < this.stageHeight) {
-            thisFigureContainer.setBaseY(this.stageHeight / 2 + this.stageHeight - targetHeight / 2);
-          }
-          if (presetPosition === 'center') {
-            thisFigureContainer.setBaseX(this.stageWidth / 2);
-          }
-          if (presetPosition === 'left') {
-            thisFigureContainer.setBaseX(targetWidth / 2);
-          }
-          if (presetPosition === 'right') {
-            thisFigureContainer.setBaseX(this.stageWidth - targetWidth / 2);
-          }
-          thisFigureContainer.pivot.set(0, this.stageHeight / 2);
-          thisFigureContainer.addChild(figureSprite);
-        }
-      }, 0);
-    };
-
-    /**
-     * 加载器部分
-     */
-    this.cacheGC();
-    if (!loader.resources?.[url]) {
-      this.loadAsset(url, setup, spineId);
     } else {
       // 复用
       setup();
@@ -931,6 +760,26 @@ export default class PixiStage {
     PIXI.utils.clearTextureCache();
   }
 
+  public getExtName(url: string) {
+    return url.split('.').pop() ?? 'png';
+  }
+
+  public getFigureMetadataByKey(key: string): IFigureMetadata | undefined {
+    console.log(key, webgalStore.getState().stage.figureMetaData);
+    return webgalStore.getState().stage.figureMetaData[key];
+  }
+
+  public loadAsset(url: string, callback: () => void, name?: string) {
+    /**
+     * Loader 复用疑似有问题，转而采用先前的单独方式
+     */
+    this.loadQueue.unshift({ url, callback, name });
+    /**
+     * 尝试启动加载
+     */
+    this.callLoader();
+  }
+
   private updateL2dMotionByKey(target: string, motion: string) {
     const figureTargetIndex = this.live2dFigureRecorder.findIndex((e) => e.target === target);
     if (figureTargetIndex >= 0) {
@@ -947,17 +796,6 @@ export default class PixiStage {
     } else {
       this.live2dFigureRecorder.push({ target, motion: '', expression });
     }
-  }
-
-  private loadAsset(url: string, callback: () => void, name?: string) {
-    /**
-     * Loader 复用疑似有问题，转而采用先前的单独方式
-     */
-    this.loadQueue.unshift({ url, callback, name });
-    /**
-     * 尝试启动加载
-     */
-    this.callLoader();
   }
 
   private callLoader() {
@@ -1005,15 +843,6 @@ export default class PixiStage {
   private unlockStageObject(targetName: string) {
     const index = this.lockTransformTarget.findIndex((name) => name === targetName);
     if (index >= 0) this.lockTransformTarget.splice(index, 1);
-  }
-
-  private getExtName(url: string) {
-    return url.split('.').pop() ?? 'png';
-  }
-
-  private getFigureMetadataByKey(key: string): IFigureMetadata | undefined {
-    console.log(key, webgalStore.getState().stage.figureMetaData);
-    return webgalStore.getState().stage.figureMetaData[key];
   }
 }
